@@ -2,7 +2,7 @@ import { Icon } from 'office-ui-fabric-react/lib/Icon';
 import * as React from 'react';
 import ReactMarkdown from 'react-markdown';
 import timeago from 'timeago.js';
-import { coerce, gt, SemVer } from 'semver';
+import { coerce, eq, gt, SemVer } from 'semver';
 
 import { config } from '../config';
 import Dependencies from './Dependencies';
@@ -27,7 +27,6 @@ interface IDisplayPackageProps {
 
 interface IPackage {
   id: string;
-  latestVersion: string;
   hasReadme: boolean;
   description: string;
   readme: string;
@@ -62,7 +61,7 @@ class DisplayPackage extends React.Component<IDisplayPackageProps, IDisplayPacka
   };
 
   private id: string;
-  private version?: string;
+  private version: SemVer | null;
 
   private registrationController: AbortController;
   private readmeController: AbortController;
@@ -74,7 +73,7 @@ class DisplayPackage extends React.Component<IDisplayPackageProps, IDisplayPacka
     this.readmeController = new AbortController();
 
     this.id = props.match.params.id.toLowerCase();
-    this.version = props.match.params.version;
+    this.version = coerce(props.match.params.version);
     this.state = DisplayPackage.initialState;
   }
 
@@ -94,7 +93,7 @@ class DisplayPackage extends React.Component<IDisplayPackageProps, IDisplayPacka
       this.readmeController = new AbortController();
 
       this.id = this.props.match.params.id.toLowerCase();
-      this.version = this.props.match.params.version;
+      this.version = coerce(this.props.match.params.version);
       this.setState(DisplayPackage.initialState);
       this.componentDidMount();
     }
@@ -126,9 +125,13 @@ class DisplayPackage extends React.Component<IDisplayPackageProps, IDisplayPacka
         if (!entry.catalogEntry.listed) continue;
 
         const normalizedVersion = this.normalizeVersion(entry.catalogEntry.version);
-        const isCurrent = !!this.version
-          ? normalizedVersion === this.version
-          : normalizedVersion === latestVersion;
+        const coercedVersion = coerce(entry.catalogEntry.version);
+
+        if (coercedVersion === null) continue;
+
+        const isCurrent = latestVersion !== null && coercedVersion !== null
+          ? eq(coercedVersion, !!this.version ? this.version : latestVersion)
+          : false;
 
         versions.push({
           date: new Date(entry.catalogEntry.published),
@@ -167,7 +170,6 @@ class DisplayPackage extends React.Component<IDisplayPackageProps, IDisplayPacka
             downloadUrl: currentItem.packageContent,
             packageType,
             lastUpdate,
-            latestVersion,
             normalizedVersion: this.normalizeVersion(currentItem.catalogEntry.version),
             readme,
             totalDownloads: results.totalDownloads,
@@ -316,8 +318,7 @@ class DisplayPackage extends React.Component<IDisplayPackageProps, IDisplayPacka
       : version.substring(0, buildMetadataStart);
   }
 
-  private latestVersion(index: Registration.IRegistrationIndex): string | null {
-    let latestVersionString: string | null = null;
+  private latestVersion(index: Registration.IRegistrationIndex): SemVer | null {
     let latestVersion: SemVer | null = null;
     for (const entry of index.items[0].items) {
       if (!entry.catalogEntry.listed) continue;
@@ -325,13 +326,12 @@ class DisplayPackage extends React.Component<IDisplayPackageProps, IDisplayPacka
       let entryVersion = coerce(entry.catalogEntry.version);
       if (!!entryVersion) {
         if (latestVersion === null || gt(entryVersion, latestVersion)) {
-          latestVersionString = entry.catalogEntry.version;
           latestVersion = entryVersion;
         }
       }
     }
 
-    return latestVersionString;
+    return latestVersion;
   }
 }
 
